Moved the shop and wishlist context into the context folder for easier use. useField hook added for future use in all fields.
