+++ Toivoa_Frontend/src/hooks/useField.js
@@ -1,0 +1,13 @@
+
+const useField = (type,value,setValue) => {
+
+  const onChange = (event) => setValue(event.target.value);
+
+  return {
+    type,
+    value,
+    onChange,
+  };
+};
+
+export default useField;
